moved databricksmanager to integrationtest
