+++ source/dotnet/Services/IntegrationTests/Components/DatabricksHttpListener.cs
@@ -18,7 +18,7 @@ using Energinet.DataHub.Wholesale.Components.DatabricksClient;
 using Microsoft.Azure.Databricks.Client;
 using Newtonsoft.Json;
 
-namespace DatabricksClientManager;
+namespace Energinet.DataHub.Wholesale.IntegrationTests.Components;
 
 /// <summary>
 /// There is currently no Databricks environment for integration testing.
