+++ source/dotnet/Services/IntegrationTests/Components/DatabricksManager.cs
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-namespace DatabricksClientManager
+namespace Energinet.DataHub.Wholesale.IntegrationTests.Components
 {
     public sealed class DatabricksManager : IAsyncDisposable
     {
