+++ source/dotnet/Services/IntegrationTests/TestCommon/Fixture/Database/ProcessManagerDatabricksFixture.cs
@@ -15,6 +15,7 @@
 using DatabricksClientManager;
 using Energinet.DataHub.Wholesale.ProcessManager;
 using Xunit;
+using DatabricksManager = Energinet.DataHub.Wholesale.IntegrationTests.Components.DatabricksManager;
 
 namespace Energinet.DataHub.Wholesale.IntegrationTests.TestCommon.Fixture.Database;
