+++ source/dotnet/Services/Wholesale.sln
@@ -34,8 +34,6 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_SolutionFolder", "_Solutio
 		stylecop.json = stylecop.json
 	EndProjectSection
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DatabricksClientManager", "DatabricksClientManager\DatabricksClientManager.csproj", "{33BF8FF4-25BA-40AA-AA43-263DF0B13B0B}"
-EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "DatabricksClient", "DatabricksClient\DatabricksClient.csproj", "{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}"
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "IntegrationTests", "IntegrationTests\IntegrationTests.csproj", "{6D00C8EA-65D9-4568-B3C6-23E31598D51E}"
@@ -90,10 +88,6 @@ Global
 		{FF7C0086-D637-4265-A667-85275B4E05EF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{FF7C0086-D637-4265-A667-85275B4E05EF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{FF7C0086-D637-4265-A667-85275B4E05EF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{33BF8FF4-25BA-40AA-AA43-263DF0B13B0B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{33BF8FF4-25BA-40AA-AA43-263DF0B13B0B}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{33BF8FF4-25BA-40AA-AA43-263DF0B13B0B}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{33BF8FF4-25BA-40AA-AA43-263DF0B13B0B}.Release|Any CPU.Build.0 = Release|Any CPU
 		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
 		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E}.Release|Any CPU.ActiveCfg = Release|Any CPU
@@ -119,7 +113,6 @@ Global
 		{FF7C0086-D637-4265-A667-85275B4E05EF} = {B6F5E0F9-7534-48E8-B4C3-719C6CE14BA0}
 		{902AE74C-5B0C-4DB2-80BB-F9A94367EB8E} = {DC447409-69A1-4705-B420-65B95EF1651E}
 		{6D00C8EA-65D9-4568-B3C6-23E31598D51E} = {A65AE4F0-E4A5-4F6C-AB74-F8A278C31AD9}
-		{33BF8FF4-25BA-40AA-AA43-263DF0B13B0B} = {DC447409-69A1-4705-B420-65B95EF1651E}
 		{EFA0CC43-7B22-4842-9AC0-E0FDA4F844BA} = {A65AE4F0-E4A5-4F6C-AB74-F8A278C31AD9}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
